added superclass
