+++ src/Component.js
@@ -1,0 +1,41 @@
+import {
+  createEvent
+} from './createEvent';
+
+class Component {
+  constructor() {
+    if (new.target === Component) {
+      throw new Error(`Can't instantiate BaseComponent, only concrete one.`);
+    }
+
+    this._element = null;
+    this._state = {};
+  }
+
+  get element() {
+    return this._element;
+  }
+
+  get template() {
+    throw new Error(`You have to define template.`);
+  }
+
+  render() {
+    this._element = createEvent(this.template);
+    this.bind();
+    return this._element;
+  }
+
+  bind() {
+
+  }
+
+  unbind() {}
+
+  unrender() {
+    this.unbind();
+    this._element = null;
+  }
+}
+
+export default Component;
